requirements.txt
alembic==1.5.6
asgiref==3.3.1
blinker==1.4
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
Django==3.1.7
dnspython==2.1.0
dominate==2.6.0
email-validator==1.1.2
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.7.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
Flask-Uploads==0.2.1
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==2.8
image==1.5.33
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==1.1.1
Pillow==8.1.2
psycopg2-binary==2.8.6
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
python-http-client==3.3.2
pytz==2021.1
requests==2.22.0
sendgrid==6.6.0
six==1.15.0
SQLAlchemy==1.3.23
sqlparse==0.4.1
starkbank-ecdsa==1.1.0
urllib3==1.25.11
visitor==0.1.3
Werkzeug==0.16.0
WTForms==2.3.3
